+++ ng-package.json
@@ -1,8 +1,7 @@
 {
   "$schema": "./node_modules/ng-packagr/ng-package.schema.json",
   "dest": "./dist",
-  "deleteDestPath": false,
   "lib": {
-    "entryFile": "./public_api.ts"
+    "entryFile": "./src/public_api.ts"
   }
 }
