--- public_api.ts
@@ -1,23 +1,0 @@
-/*
- * Public API Surface of ng-http-loader
- */
-
-export * from './src/components/sk-chasing-dots/sk-chasing-dots.component';
-export * from './src/components/sk-cube-grid/sk-cube-grid.component';
-export * from './src/components/sk-double-bounce/sk-double-bounce.component';
-export * from './src/components/sk-rotating-plane/sk-rotating-plane.component';
-export * from './src/components/sk-spinner-pulse/sk-spinner-pulse.component';
-export * from './src/components/sk-three-bounce/sk-three-bounce.component';
-export * from './src/components/sk-wandering-cubes/sk-wandering-cubes.component';
-export * from './src/components/sk-wave/sk-wave.component';
-export * from './src/components/spinner/spinner.component';
-
-export * from './src/components/abstract.loader.component';
-export * from './src/components/ng-http-loader-components.module';
-
-export * from './src/services/ng-http-loader-services.module';
-export * from './src/services/pending-interceptor.service';
-export * from './src/services/spinner-visibility.service';
-
-export * from './src/ng-http-loader.module';
-export * from './src/spinkits';
