+++ src/lib/components/sk-chasing-dots/sk-chasing-dots.component.ts
