+++ src/lib/components/sk-double-bounce/sk-double-bounce.component.ts
