+++ src/lib/components/sk-rotating-plane/sk-rotating-plane.component.css
