+++ src/lib/components/sk-wave/sk-wave.component.ts
