+++ src/lib/ng-http-loader.module.ts
