+++ src/lib/spinkits.ts
