+++ src/test/components/spinner/spinner.outlet.component.spec.ts
@@ -11,10 +11,10 @@ import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { BrowserDynamicTestingModule } from '@angular/platform-browser-dynamic/testing';
-import { SkThreeBounceComponent } from '../../../src/components/sk-three-bounce/sk-three-bounce.component';
-import { SpinnerComponent } from '../../../src/components/spinner/spinner.component';
-import { NgHttpLoaderServicesModule } from '../../../src/services/ng-http-loader-services.module';
-import { SPINKIT_COMPONENTS } from '../../../src/spinkits';
+import { SkThreeBounceComponent } from '../../../lib/components/sk-three-bounce/sk-three-bounce.component';
+import { SpinnerComponent } from '../../../lib/components/spinner/spinner.component';
+import { NgHttpLoaderServicesModule } from '../../../lib/services/ng-http-loader-services.module';
+import { SPINKIT_COMPONENTS } from '../../../lib/spinkits';
 
 describe('SpinnerNgOutletComponent', () => {
     let component: SpinnerComponent;
