+++ src/test/services/spinner-visibility.service.spec.ts
@@ -8,8 +8,8 @@
  */
 
 import { inject, TestBed } from '@angular/core/testing';
-import { PendingInterceptorService } from '../../src/services/pending-interceptor.service';
-import { SpinnerVisibilityService } from '../../src/services/spinner-visibility.service';
+import { PendingInterceptorService } from '../../lib/services/pending-interceptor.service';
+import { SpinnerVisibilityService } from '../../lib/services/spinner-visibility.service';
 
 describe('SpinnerVisibilityService', () => {
