+++ tsconfig.lib.json
@@ -25,5 +25,9 @@
     "strictInjectionParameters": true,
     "flatModuleId": "AUTOGENERATED",
     "flatModuleOutFile": "AUTOGENERATED"
-  }
+  },
+  "exclude": [
+    "./src/test.ts",
+    "**/*.spec.ts"
+  ]
 }
